// src/app/kimios-client-api/model/dMEntityWrapper.ts
/**
 * Kimios API
 * Kimios Sample API
 *
 * OpenAPI spec version: 1.3-SNAPSHOT - e93001be88cfd78cb8424501a26a7860f7c03847
 * Contact: documentation@kimios.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { DMEntity } from './dMEntity';


export interface DMEntityWrapper { 
    dmEntity?: DMEntity;
    canRead?: boolean;
    canWrite?: boolean;
    hasFullAccess?: boolean;
}